+++ kernel/app/src/js/plugins/popover/index.js
@@ -1,4 +1,6 @@
 import { Base } from './../base/index.js';
+import { drag } from './drag.js';
+import { open } from './open.js';
 
 export {
     Popover as Plugin,
@@ -154,113 +156,17 @@ class Popover extends Base {
                     <div class="block__ew"></div>
                     <div class="block__ns"></div>`;
 
-                    /* 拖动功能 */
-                    const border = block__popover.querySelector(".block__icons--border");
                     const iframe = block__popover.querySelector("iframe");
+                    const border = block__popover.querySelector(".block__icons--border");
+                    const icon_pin = block__popover.querySelector('[data-type="pin"]');
+                    const icon_close = block__popover.querySelector('[data-type="close"]');
+                    const icon_open_page = block__popover.querySelector('[data-type="open-page"]');
+                    const icon_open_window = block__popover.querySelector('[data-type="open-window"]');
+                    const size_handle = block__popover.querySelector(".block__nwse");
                     const width_handle = block__popover.querySelector(".block__ew");
                     const height_handle = block__popover.querySelector(".block__ns");
-                    const size_handle = block__popover.querySelector(".block__nwse");
-
-                    // 鼠标移动时状态
-                    var flag_popover_dragging = false; // 悬浮预览窗口是否正在拖动
-
-                    /* 标题栏拖动功能 */
-                    /* 窗口拖动注册 */
-                    const that = this;
-                    function dragRegister(element, mousemoveHandler) {
-                        element.addEventListener("mousedown", e => {
-                            /* 取消其他默认事件处理 */
-                            e.preventDefault();
-                            e.stopPropagation();
-
-                            flag_popover_dragging = true; // 正在拖拽
-                            block__popover.style.zIndex = popover.z_index++; // 将当前窗口置顶
-                            /* 记录鼠标与窗口的相对位置 */
-                            popover.drag.position.x = e.clientX - block__popover.offsetLeft; // 鼠标相对于子窗口左上角的横向偏移量(鼠标横坐标 - popover 的 左侧偏移量)
-                            popover.drag.position.y = e.clientY - block__popover.offsetTop; // 鼠标相对于子窗口左上角的纵向偏移量(鼠标纵坐标 - popover 的 上侧偏移量)
-                            popover.drag.size.width = block__popover.offsetWidth; // 窗口宽度
-                            popover.drag.size.height = block__popover.offsetHeight; // 窗口高度
-
-                            that.dragMousedown(e, doc, mousemoveHandler);
-                        });
-                        element.addEventListener("mouseup", e => { // 松开按键结束拖拽
-                            /* 取消其他默认事件处理 */
-                            e.preventDefault();
-                            e.stopPropagation();
-
-                            flag_popover_dragging = false;
-                            that.dragMouseup(e, doc, mousemoveHandler);
-                        });
-                        doc.addEventListener('mouseleave', e => { // 鼠标移出拖拽区域结束拖拽
-                            flag_popover_dragging = false;
-                            this.dragMouseup(e, doc, eventHandler);
-                        }, false);
-                    }
-                    // REF [JS拖动浮动DIV - boystar - 博客园](https://www.cnblogs.com/boystar/p/5231697.html)
-                    // REF [JS鼠标事件完成元素拖拽（简单-高级） - 百度文库](https://wenku.baidu.com/view/0c56050c3269a45177232f60ddccda38376be161?bfetype=new)
-                    /* 悬浮子窗口拖拽 */
-                    function popoverDrag(e) {
-                        // console.log(e);
-                        /* 子窗口左上角将要移动到的位置坐标 */
-                        let x = e.clientX - popover.drag.position.x;
-                        let y = e.clientY - popover.drag.position.y;
-
-                        /* 子窗口左上角可以可以移动到区域边缘 */
-                        let window_width = doc.documentElement.clientWidth - block__popover.offsetWidth;
-                        let window_height = doc.documentElement.clientHeight - block__popover.offsetHeight;
-
-                        x = (x < 0) ? 0 : x;                          // 当子窗口移动到主窗口最左边时
-                        x = (x > window_width) ? window_width : x;    // 当子窗口移动到主窗口最右边时
-                        y = (y < 0) ? 0 : y;                          // 当子窗口移动到主窗口最上边时
-                        y = (y > window_height) ? window_height : y;  // 当子窗口移动到主窗口最下边时
-
-                        block__popover.style.left = `${x}px`;
-                        block__popover.style.top = `${y}px`;
-                    }
-
-                    /* 子窗口尺寸拖动调整功能 */
-                    function calcSize(e) { // 计算子窗口应该调整到的尺寸
-                        /** 计算窗口将要调整到的宽度
-                         *  鼠标当前位置(应优化为控件的中轴) - 窗口左边位置
-                         *      = 鼠标相对于窗口左边的横向偏移量
-                         *  鼠标相对于窗口左边的横向偏移量 - 拖动前鼠标相对于窗口左边的横向偏移量
-                         *      = 鼠标宽度变化量
-                         *  原宽度 - 鼠标宽度变化量
-                         *      = 当前应当设置的宽度
-                         */
-                        let width = popover.drag.size.width + (e.clientX - block__popover.offsetLeft - popover.drag.position.x);
-                        /* 窗口宽度上限: 页面可视宽度 - 窗口左边位置 */
-                        let max_width = doc.documentElement.clientWidth - block__popover.offsetLeft;
-                        width = (width < 0) ? 0 : width; // 当子窗口移动到主窗口最左边时
-                        width = (width > max_width) ? max_width : width; // 当子窗口移动到主窗口最右边时
-
-                        let height = popover.drag.size.height + (e.clientY - block__popover.offsetTop - popover.drag.position.y);
-                        let max_height = doc.documentElement.clientHeight - block__popover.offsetTop;
-                        height = (height < 0) ? 0 : height; // 当子窗口移动到主窗口最上边时
-                        height = (height > max_height) ? max_height : height; // 当子窗口移动到主窗口最下边时
-                        return {
-                            width: width,
-                            height: height,
-                        };
-                    }
-                    function widthDrag(e) {
-                        block__popover.style.width = `${calcSize(e).width}px`;
-                    }
-                    function heightDrag(e) {
-                        block__popover.style.height = `${calcSize(e).height}px`;
-                    }
-                    function sizeDrag(e) {
-                        const size = calcSize(e);
-                        block__popover.style.width = `${size.width}px`;
-                        block__popover.style.height = `${size.height}px`;
-                    }
-                    dragRegister(border, popoverDrag);
-                    dragRegister(width_handle, widthDrag);
-                    dragRegister(height_handle, heightDrag);
-                    dragRegister(size_handle, sizeDrag);
 
                     /* 鼠标移出预览时关闭预览 */
-                    const icon_pin = block__popover.querySelector('[data-type="pin"]');
                     function close(_) {
                         if (!flag_popover_dragging && !icon_pin.classList.contains("block__icon--active")) {
                             setTimeout(() => block__popover.remove(), TIMEOUT.CLOSE);
@@ -281,27 +187,117 @@ class Popover extends Base {
                     border.addEventListener("dblclick", pin); // 标题栏双击
 
                     /* 关闭按钮单击 */
-                    block__popover.querySelector('[data-type="close"]').addEventListener("click", _ => block__popover.remove());
+                    icon_close.addEventListener("click", _ => block__popover.remove());
                     /* 鼠标移出元素后自动关闭 */
                     setTimeout(() => block__popover.addEventListener("mouseleave", close), TIMEOUT.SHOW_MIN);
 
                     /* 在新窗口打开 */
                     // REF [Window.open() - Web APIs | MDN](https://developer.mozilla.org/en-US/docs/Web/API/Window/open)
                     // REF [Window open() 方法 | 菜鸟教程](https://www.runoob.com/jsref/met-win-open.html)
-                    const icon_open_page = block__popover.querySelector('[data-type="open-page"]');
-                    const icon_open_window = block__popover.querySelector('[data-type="open-window"]');
                     icon_open_page.addEventListener("click", _ => this.context.top.open(iframe.src));
-                    icon_open_window.addEventListener("click", _ => this.context.top.open(
-                        iframe.src,
-                        iframe.src,
-                        `
-                        popup = true,
-                        left = ${this.context.top.screenX + block__popover.offsetLeft},
-                        top = ${this.context.top.screenY + block__popover.offsetTop},
-                        width = ${block__popover.offsetWidth},
-                        height = ${block__popover.offsetHeight},
-                    `,
-                    ));
+                    icon_open_window.addEventListener("click", e => {
+                        // console.log(e);
+                        // console.log(e.target.getBoundingClientRect());
+                        // console.log(block__popover.getBoundingClientRect());
+                        const position = block__popover.getBoundingClientRect();
+                        const x_relative = e.x - position.x;
+                        const y_relative = e.y - position.y;
+                        const left = e.screenX - x_relative;
+                        const top = e.screenY - y_relative;
+                        let width;
+                        let height;
+                        if (this.context.top.navigator.userAgent.indexOf("Firefox") > -1) {
+                            width = iframe.offsetWidth;
+                            height = iframe.offsetHeight;
+                        }
+                        else {
+                            width = block__popover.offsetWidth;
+                            height = block__popover.offsetHeight;
+                        }
+                        open(
+                            iframe.src,
+                            iframe.src,
+                            // this.context.top.screenX + block__popover.offsetLeft,
+                            // this.context.top.screenY + block__popover.offsetTop,
+                            left,
+                            top,
+                            width,
+                            height,
+                        );
+                    });
+
+                    /* 拖动功能 */
+
+                    // 鼠标移动时状态
+                    var flag_popover_dragging = false; // 悬浮预览窗口是否正在拖动
+
+                    /* 标题栏拖动功能 */
+                    /* 窗口拖动注册 */
+                    // REF [JS拖动浮动DIV - boystar - 博客园](https://www.cnblogs.com/boystar/p/5231697.html)
+                    // REF [JS鼠标事件完成元素拖拽（简单-高级） - 百度文库](https://wenku.baidu.com/view/0c56050c3269a45177232f60ddccda38376be161?bfetype=new)
+
+                    drag.register(
+                        border,
+                        block__popover,
+                        doc.documentElement,
+                        drag.handler.move,
+                        undefined,
+                        undefined,
+                        (e, that, ..._) => {
+                            // 移出浏览器则从新窗口打开
+                            const positon = doc.documentElement.getBoundingClientRect();
+                            if (e.x < positon.left
+                                || e.x > positon.right
+                                || e.y < positon.top
+                                || e.y > positon.bottom
+                            ) {
+                                const left = e.screenX - that.status.drag.position.x;
+                                const top = e.screenY - that.status.drag.position.y;
+                                let width;
+                                let height;
+                                if (this.context.top.navigator.userAgent.indexOf("Firefox") > -1) {
+                                    width = iframe.offsetWidth;
+                                    height = iframe.offsetHeight;
+                                }
+                                else {
+                                    width = block__popover.offsetWidth;
+                                    height = block__popover.offsetHeight;
+                                }
+                                open(
+                                    iframe.src,
+                                    iframe.src,
+                                    // this.context.top.screenX + block__popover.offsetLeft,
+                                    // this.context.top.screenY + block__popover.offsetTop,
+                                    left,
+                                    top,
+                                    width,
+                                    height,
+                                );
+                                if (!icon_pin.classList.contains("block__icon--active")) {
+                                    /* 浮窗未被钉住, 则关闭 */
+                                    icon_close.click();
+                                }
+                            }
+                        },
+                    );
+                    drag.register(
+                        size_handle,
+                        block__popover,
+                        doc.documentElement,
+                        drag.handler.resize,
+                    );
+                    drag.register(
+                        width_handle,
+                        block__popover,
+                        doc.documentElement,
+                        drag.handler.rewidth,
+                    );
+                    drag.register(
+                        height_handle,
+                        block__popover,
+                        doc.documentElement,
+                        drag.handler.reheight,
+                    );
 
                     doc.body.append(block__popover);
                 }
@@ -338,7 +334,7 @@ class Popover extends Base {
      * @params {HTMLElement} stage: 在哪个元素内拖拽
      * @params {function} mousemoveHandler: 鼠标移动事件的处理器
      */
-    dragMousedown(e, stage, mousemoveHandler) {
+    dragStart(e, stage, mousemoveHandler) {
         /* 取消其他默认事件处理 */
         e.preventDefault();
         e.stopPropagation();