+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - 修复无法划选文本问题 | Fixed an issue where text could not be sized.
 - 修复静态资源请求错误时创建空文件问题 | Fixed an issue where an empty file was created when a static resource request was wrong.
 - 鼠标移出窗口时结束拖拽 | End dragging when the mouse moves out of the window.
+- 优化鼠标移出窗口时结束拖拽 | Optimized to end dragging when the mouse moves out of the window.
 
 ## v0.1.3 / 2022-09-15